requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
requests==2.28.2
elasticsearch==8.7.0
Flask-WTF==1.1.1
pytest==7.4.2
pytest-flask==1.2.0
gunicorn==21.2.0
Flask-Login==0.6.2
Werkzeug==2.2.3
Flask-Migrate==4.0.4
Pillow==9.5.0
tqdm==4.65.0 